Move the expression validation into Velox backend.
