+++ gluten-core/src/main/scala/io/glutenproject/expression/ExpressionConverter.scala
@@ -17,10 +17,8 @@
 
 package io.glutenproject.expression
 
-import io.glutenproject.GlutenConfig
 import io.glutenproject.backendsapi.BackendsApiManager
 import io.glutenproject.execution.{NativeColumnarToRowExec, WholeStageTransformerExec}
-import io.glutenproject.expression.ConverterUtils.VELOX_EXPR_BLACKLIST
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.execution._
@@ -29,38 +27,9 @@ import org.apache.spark.sql.types.DecimalType
 
 object ExpressionConverter extends Logging {
 
-  /**
-   * Add the validation for Velox unsupported or mismatched expressions with specific input type,
-   * such as Cast(ArrayType).
-   */
-  def doValidate(blacklist: Map[String, String], expr: Expression): Boolean = {
-    val value = blacklist.get(expr.prettyName.toLowerCase())
-    if (value.isEmpty) {
-      return true
-    }
-    val inputTypeName = value.get
-    if (inputTypeName.equals("")) {
-      return false
-    } else {
-      for (input <- expr.children) {
-        if (inputTypeName.equals(input.dataType.typeName)) {
-          return false
-        }
-      }
-    }
-    true
-  }
-
-  /**
-   * Do validate the expressions based on the specific backend blacklist,
-   * the existed expression will fall back to Vanilla Spark.
-   */
-  def doValidate(expr: Expression): Boolean =
-    GlutenConfig.getSessionConf.isVeloxBackend && doValidate(VELOX_EXPR_BLACKLIST, expr)
-
   def replaceWithExpressionTransformer(expr: Expression,
       attributeSeq: Seq[Attribute]): Expression = {
-    if (!doValidate(expr)) {
+    if (!BackendsApiManager.getTransformerApiInstance.doValidate(expr)) {
       throw new UnsupportedOperationException(s"Not supported: $expr.")
     }
     expr match {
